+++ src/irclog2html/irclog2html.py
@@ -363,17 +363,25 @@ class AbstractStyle(object):
     def __init__(self, outfile, colours=None):
         """Create a text formatter for writing to outfile.
 
-        `colours` may have the following attributes:
-           part
-           join
-           server
-           nickchange
-           action
+        The ``colours`` dictionary may have the following items:
+
+        - part
+        - join
+        - server
+        - nickchange
+        - action
+
         """
         self.outfile = io.TextIOWrapper(outfile, encoding=self.charset,
                                         errors='xmlcharrefreplace',
                                         line_buffering=True)
         self.colours = colours or {}
+
+    def __del__(self):
+        """Destructor to make sure we don't close outfile prematurely."""
+        if not self.outfile.closed:
+            self.outfile.flush()
+            self.outfile.detach()  # don't let TextIOWrapper.__del__ close it!
 
     def head(self, title, prev=('', ''), index=('', ''), next=('', ''),
              searchbox=False):