+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - irclogsearch can now be deployed via WSGI.  Contributed by Albertas Agejevas
   (https://github.com/mgedmin/irclog2html/pull/9).
 
+- There's a new ``irclogserver`` script that can be used to serve
+  dynamically-generated IRC logs.  It can also be deployed via WSGI.
+
 - Drop support for Python 2.6.